Game now finishes when it is supposed to
That means that everything works for agents to play the game!  There's
a lot of beautification left to do, both in the code and in the logging
(like, we don't even say who wins at the end yet), but at least things
work.  Next up: write a web client so humans can play the game, and make
sure things work there.  Then I'll beautify everything and make it really
easy to write additional agents.